Add base Model

+++ app/Utils/Session.php
@@ -31,8 +31,11 @@ class Session implements SessionHandlerInterface
     public function read($session_id)
     {
         // Read session data and return it
-        $stmt = $this->db->prepare("SELECT * FROM sessions WHERE sid=? AND ip_address=?");
-        $stmt->execute([$session_id, $_SERVER['REMOTE_ADDR']]);
+        $stmt = $this->db->prepare("SELECT * FROM sessions WHERE sid = :sid AND ip_address = :ip_address");
+        $stmt->execute([
+            'sid' => $session_id,
+            'ip_address' => $_SERVER['REMOTE_ADDR']
+        ]);
 
         $session = $stmt->fetch(PDO::FETCH_ASSOC);
 
@@ -55,8 +58,8 @@ class Session implements SessionHandlerInterface
 
             if ($user_id != 0) {
                 // If there is already a session
-                $stmt = $this->db->prepare("SELECT sid FROM sessions WHERE user_id=?");
-                $stmt->execute([$user_id]);
+                $stmt = $this->db->prepare("SELECT sid FROM sessions WHERE user_id = :user_id");
+                $stmt->execute(['user_id' => $user_id]);
                 $num_rows = $stmt->rowCount();
 
                 if ($num_rows != 0) {
@@ -65,19 +68,30 @@ class Session implements SessionHandlerInterface
 
                     if ($row['session_id'] != $session_id) {
 
-                        $stmt = $this->db->prepare("DELETE FROM sessions WHERE user_id=?");
-                        $stmt->execute([$user_id]);
+                        $stmt = $this->db->prepare("DELETE FROM sessions WHERE user_id = :user_id");
+                        $stmt->execute(['user_id' => $user_id]);
                     }
                 }
             }
 
-            $stmt = $this->db->prepare("INSERT INTO sessions (sid, user_id, session_data, ip_address, created_on, updated_on) VALUES (?,?,?,?,CURRENT_TIMESTAMP,CURRENT_TIMESTAMP)");
-            $stmt->execute([$session_id, $user_id, $session_data, $ip_address]);
+            $stmt = $this->db->prepare("INSERT INTO sessions (sid, user_id, session_data, ip_address, created_at, updated_at)
+                                        VALUES (:sid, :user_id, :session_data, :ip_address, CURRENT_TIMESTAMP,CURRENT_TIMESTAMP)");
+            $stmt->execute([
+                'sid' => $session_id,
+                'user_id' => $user_id,
+                'session_data' => $session_data,
+                'ip_address' => $ip_address
+            ]);
 
         } else {
             // If the session exists in the database update session data and last_updated
-            $stmt = $this->db->prepare("UPDATE sessions SET session_data=?, updated_on=CURRENT_TIMESTAMP WHERE sid=? AND ip_address=? AND user_id=?");
-            $stmt->execute([$session_data, $session_id, $ip_address, $user_id]);
+            $stmt = $this->db->prepare("UPDATE sessions SET session_data = :session_data, updated_at = CURRENT_TIMESTAMP WHERE sid = :sid AND ip_address = :ip_address AND user_id = :user_id");
+            $stmt->execute([
+                'session_data' => $session_data,
+                'sid' => $session_id,
+                'ip_address' => $ip_address,
+                'user_id' => $user_id
+            ]);
         }
 
         return true;
@@ -87,13 +101,13 @@ class Session implements SessionHandlerInterface
     public function destroy($session_id)
     {
         // Destroy the session
-        $stmt = $this->db->prepare("SELECT user_id FROM sessions WHERE sid=?");
-        $stmt->execute([$session_id]);
+        $stmt = $this->db->prepare("SELECT user_id FROM sessions WHERE sid = :sid");
+        $stmt->execute(['sid' => $session_id]);
         $row = $stmt->fetch(PDO::FETCH_ASSOC);
 
         if (is_array($row)) {
-            $stmt = $this->db->prepare("DELETE FROM sessions WHERE sid=?");
-            $stmt->execute([$session_id]);
+            $stmt = $this->db->prepare("DELETE FROM sessions WHERE sid = :sid");
+            $stmt->execute(['sid' => $session_id]);
         }
 
         return true;
@@ -101,7 +115,7 @@ class Session implements SessionHandlerInterface
 
     public function gc($maxlifetime = 86400) // One day
     {
-        $stmt = $this->db->prepare("DELETE FROM sessions WHERE updated_on < CURRENT_TIMESTAMP - INTERVAL '$maxlifetime' SECOND");
+        $stmt = $this->db->prepare("DELETE FROM sessions WHERE updated_at < CURRENT_TIMESTAMP - INTERVAL '$maxlifetime' SECOND");
 
         $stmt->execute();
